TweetAnalyzerwithLLM/requirements.txt
flask==2.3.3
werkzeug==2.3.7
google-generativeai==0.3.2
nest_asyncio